+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **[IMPROVEMENT]** Reduced motion now applies to most of hoverable UI elements
 * **[FIX]** Reset idea create modal description when posting new idea (reported at [Feedbacky](https://app.feedbacky.net/i/reset-description-field-on-new-idea.4045))
 * **[FIX]** `and X more` message didn't appear properly
+* **[FIX]** Changelog create modal didn't preserve description on close (reported at [Feedbacky](https://app.feedbacky.net/i/changelog-create-modal-issue.4075))
 
 ### 1.0.0.RC.0
 * **[FEATURE]** Reactions will now display who reacted